fix(release): close Alpha.3 source acceptance
