+++ framework/tui/src/main.tsx
@@ -153,11 +153,21 @@ import {
 const nodeRequire = createRequire(import.meta.url);
 let tuiAgentSessionEndpoint = '';
 let tuiAgentSessionReady: Promise<string> | undefined;
-let tuiAgentSessionHost:
-  | ReturnType<typeof createAttachedAgentSessionHost>
-  | ReturnType<typeof createDetachedAgentSessionHost>
-  | undefined;
+type TuiAgentSessionHost = ReturnType<typeof createAttachedAgentSessionHost>;
+let tuiAgentSessionHost: TuiAgentSessionHost | undefined;
 let tuiAgentSessionRuntimeDir = '';
+function tuiAgentSessionEnvironment(
+  env: NodeJS.ProcessEnv,
+  optional = false,
+): NodeJS.ProcessEnv {
+  if (optional && !tuiAgentSessionEndpoint) return { ...env };
+  return {
+    ...env,
+    KF_RUNTIME_DIR: tuiAgentSessionRuntimeDir,
+    KUNGFU_AGENT_SESSION_ENDPOINT: tuiAgentSessionEndpoint,
+    KUNGFU_PROJECT_WORK_AGENT_SESSION: '1',
+  };
+}
 function ensureTuiAgentSession(runtimeDir: string): Promise<string> {
   const resolvedRuntimeDir = path.resolve(runtimeDir);
   if (
@@ -177,20 +187,10 @@ function ensureTuiAgentSession(runtimeDir: string): Promise<string> {
     argvEntry: process.argv[1],
     modulePath: fileURLToPath(import.meta.url),
   });
-  const packagedPty = path.join(
-    path.dirname(workerPath),
-    'node_modules',
-    'node-pty',
-    'lib',
-    'index.js',
-  );
-  const sourcePty = path.join(
-    packageRoot,
-    'node_modules',
-    'node-pty',
-    'lib',
-    'index.js',
-  );
+  const ptyModule = (root: string) =>
+    path.join(root, 'node_modules', 'node-pty', 'lib', 'index.js');
+  const packagedPty = ptyModule(path.dirname(workerPath));
+  const sourcePty = ptyModule(packageRoot);
   process.env.KUNGFU_AGENT_SESSION_NODE_PTY_MODULE = prepareAgentSessionNodePty(
     {
       runtimeDir: resolvedRuntimeDir,
@@ -201,9 +201,8 @@ function ensureTuiAgentSession(runtimeDir: string): Promise<string> {
   process.env.KUNGFU_MOCK_AGENT_SCRIPT = mockPath;
   process.env.KUNGFU_PROJECT_WORK_AGENT_SESSION = '1';
   const paths = runtimePaths();
-  const deterministicMock = Boolean(
-    process.env.KUNGFU_MOCK_AGENT_SCENARIO?.trim(),
-  );
+  const mockScenario = process.env.KUNGFU_MOCK_AGENT_SCENARIO;
+  const deterministicMock = Boolean(mockScenario?.trim());
   const host = deterministicMock
     ? createAttachedAgentSessionHost({
         runtimeDir: resolvedRuntimeDir,
@@ -234,8 +233,7 @@ async function closeTuiAgentSession() {
   const host = tuiAgentSessionHost;
   tuiAgentSessionHost = undefined;
   tuiAgentSessionReady = undefined;
-  tuiAgentSessionEndpoint = '';
-  tuiAgentSessionRuntimeDir = '';
+  tuiAgentSessionEndpoint = tuiAgentSessionRuntimeDir = '';
   if (host && 'close' in host) await host.close();
 }
 async function invokeTuiAgentSession(
@@ -303,6 +301,95 @@ function tuiCliInvocation(paths: ReturnType<typeof runtimePaths>) {
     args: (values: string[]) => [...argsPrefix, ...values],
   };
 }
+function bindMockAgentEnvironment(
+  cli: ReturnType<typeof tuiCliInvocation>,
+  paths: ReturnType<typeof runtimePaths>,
+) {
+  const { mockPath } = resolveTuiAgentSessionPaths({
+    env: process.env,
+    argvEntry: process.argv[1],
+    modulePath: fileURLToPath(import.meta.url),
+  });
+  cli.env = bindTuiMockAgentEnvironment({
+    env: cli.env,
+    packagedBin: paths.packagedBin,
+    mockPath,
+  });
+}
+function streamCliEvents({
+  file,
+  args,
+  env,
+  maxBuffer,
+  onLine,
+  label,
+}: {
+  file: string;
+  args: string[];
+  env: NodeJS.ProcessEnv;
+  maxBuffer: number;
+  onLine: (line: string) => void;
+  label: string;
+}): Promise<void> {
+  return new Promise((resolve, reject) => {
+    const child = spawn(file, args, {
+      env,
+      stdio: ['ignore', 'pipe', 'pipe'],
+    });
+    let stdoutBuffer = '';
+    const stderrChunks: unknown[] = [];
+    let outputSize = 0;
+    let settled = false;
+    const fail = (error: Error) => {
+      if (settled) return;
+      settled = true;
+      child.kill();
+      reject(error);
+    };
+    const emitLine = (line: string) => {
+      if (!line.trim()) return true;
+      try {
+        onLine(line);
+        return true;
+      } catch (reason) {
+        fail(
+          reason instanceof Error
+            ? reason
+            : new Error(`invalid ${label} event: ${String(reason)}`),
+        );
+        return false;
+      }
+    };
+    child.stdout.on('data', (chunk) => {
+      const text = String(chunk);
+      outputSize += text.length;
+      if (outputSize > maxBuffer) {
+        fail(new Error(`${label} event stream exceeded maxBuffer`));
+        return;
+      }
+      stdoutBuffer += text;
+      const lines = stdoutBuffer.split(/\r?\n/u);
+      stdoutBuffer = lines.pop() ?? '';
+      for (const line of lines) if (!emitLine(line)) return;
+    });
+    child.stderr.on('data', stderrChunks.push.bind(stderrChunks));
+    child.once('error', fail);
+    child.once('close', (code, signal) => {
+      if (settled || !emitLine(stdoutBuffer)) return;
+      if (code !== 0) {
+        fail(
+          new Error(
+            stderrChunks.join('').trim() ||
+              `${label} event stream exited ${code ?? signal ?? 'unknown'}`,
+          ),
+        );
+        return;
+      }
+      settled = true;
+      resolve();
+    });
+  });
+}
 type ExitHistoryStatus = {
   ok: boolean;
   state: string;
@@ -320,16 +407,7 @@ const EXIT_HISTORY_STATUS_FALLBACK: ExitHistoryStatus = {
 async function openTuiAgentWorkLab(projectTour = false): Promise<AgentWorkLab> {
   const paths = runtimePaths();
   const cli = tuiCliInvocation(paths);
-  const { mockPath } = resolveTuiAgentSessionPaths({
-    env: process.env,
-    argvEntry: process.argv[1],
-    modulePath: fileURLToPath(import.meta.url),
-  });
-  cli.env = bindTuiMockAgentEnvironment({
-    env: cli.env,
-    packagedBin: paths.packagedBin,
-    mockPath,
-  });
+  bindMockAgentEnvironment(cli, paths);
   if (projectTour) {
     const endpoint = await ensureTuiAgentSession(paths.runtimeDir);
     cli.env.KUNGFU_AGENT_SESSION_ENDPOINT = endpoint;
@@ -354,77 +432,13 @@ async function openTuiAgentWorkLab(projectTour = false): Promise<AgentWorkLab> {
       if (tuiAgentSessionReady) {
         await ensureTuiAgentSession(tuiAgentSessionRuntimeDir);
       }
-      return new Promise<void>((resolve, reject) => {
-        const child = spawn(file, cli.args(values), {
-          env: {
-            ...options.env,
-            ...(tuiAgentSessionEndpoint
-              ? {
-                  KF_RUNTIME_DIR: tuiAgentSessionRuntimeDir,
-                  KUNGFU_AGENT_SESSION_ENDPOINT: tuiAgentSessionEndpoint,
-                  KUNGFU_PROJECT_WORK_AGENT_SESSION: '1',
-                }
-              : {}),
-          },
-          stdio: ['ignore', 'pipe', 'pipe'],
-        });
-        let stdoutBuffer = '';
-        let stderr = '';
-        let outputSize = 0;
-        let settled = false;
-        const fail = (error: Error) => {
-          if (settled) return;
-          settled = true;
-          child.kill();
-          reject(error);
-        };
-        const emitLine = (line: string) => {
-          if (!line.trim()) return true;
-          try {
-            onLine(line);
-            return true;
-          } catch (reason) {
-            fail(
-              reason instanceof Error
-                ? reason
-                : new Error(`invalid qualification event: ${String(reason)}`),
-            );
-            return false;
-          }
-        };
-        child.stdout.on('data', (chunk) => {
-          const text = String(chunk);
-          outputSize += text.length;
-          if (outputSize > options.maxBuffer) {
-            fail(new Error('qualification event stream exceeded maxBuffer'));
-            return;
-          }
-          stdoutBuffer += text;
-          const lines = stdoutBuffer.split(/\r?\n/);
-          stdoutBuffer = lines.pop() ?? '';
-          for (const line of lines) {
-            if (!emitLine(line)) return;
-          }
-        });
-        child.stderr.on('data', (chunk) => {
-          stderr += String(chunk);
-        });
-        child.once('error', fail);
-        child.once('close', (code, signal) => {
-          if (settled) return;
-          if (!emitLine(stdoutBuffer)) return;
-          if (code !== 0) {
-            fail(
-              new Error(
-                stderr.trim() ||
-                  `qualification event stream exited ${code ?? signal ?? 'unknown'}`,
-              ),
-            );
-            return;
-          }
-          settled = true;
-          resolve();
-        });
+      return streamCliEvents({
+        file,
+        args: cli.args(values),
+        env: tuiAgentSessionEnvironment(options.env, true),
+        maxBuffer: options.maxBuffer,
+        onLine,
+        label: 'qualification',
       });
     },
   });
@@ -432,16 +446,7 @@ async function openTuiAgentWorkLab(projectTour = false): Promise<AgentWorkLab> {
 function openTuiProjects(useAgentSession = true, allowForeignBinding = false) {
   const paths = runtimePaths();
   const cli = tuiCliInvocation(paths);
-  const { mockPath } = resolveTuiAgentSessionPaths({
-    env: process.env,
-    argvEntry: process.argv[1],
-    modulePath: fileURLToPath(import.meta.url),
-  });
-  cli.env = bindTuiMockAgentEnvironment({
-    env: cli.env,
-    packagedBin: paths.packagedBin,
-    mockPath,
-  });
+  bindMockAgentEnvironment(cli, paths);
   if (!useAgentSession) {
     cli.env.KUNGFU_PROJECT_WORK_AGENT_SESSION = undefined;
     cli.env.KUNGFU_AGENT_SESSION_ENDPOINT = undefined;
@@ -471,11 +476,7 @@ function openTuiProjects(useAgentSession = true, allowForeignBinding = false) {
           cli.args(values),
           {
             ...options,
-            env: {
-              ...options.env,
-              KF_RUNTIME_DIR: tuiAgentSessionRuntimeDir,
-              KUNGFU_AGENT_SESSION_ENDPOINT: tuiAgentSessionEndpoint,
-            },
+            env: tuiAgentSessionEnvironment(options.env),
           },
           (error, stdout, stderr) => {
             if (error)
@@ -487,11 +488,7 @@ function openTuiProjects(useAgentSession = true, allowForeignBinding = false) {
     execFileInput: (file, values, input, options) =>
       new Promise<string>((resolve, reject) => {
         const child = spawn(file, cli.args(values), {
-          env: {
-            ...options.env,
-            KF_RUNTIME_DIR: tuiAgentSessionRuntimeDir,
-            KUNGFU_AGENT_SESSION_ENDPOINT: tuiAgentSessionEndpoint,
-          },
+          env: tuiAgentSessionEnvironment(options.env),
           stdio: ['pipe', 'pipe', 'pipe'],
         });
         let stdout = '';
@@ -538,72 +535,13 @@ function openTuiProjects(useAgentSession = true, allowForeignBinding = false) {
       }),
     execFileEvents: async (file, values, options, onLine) => {
       if (tuiAgentSessionReady) await tuiAgentSessionReady;
-      return new Promise<void>((resolve, reject) => {
-        const child = spawn(file, cli.args(values), {
-          env: {
-            ...options.env,
-            KF_RUNTIME_DIR: tuiAgentSessionRuntimeDir,
-            KUNGFU_AGENT_SESSION_ENDPOINT: tuiAgentSessionEndpoint,
-          },
-          stdio: ['ignore', 'pipe', 'pipe'],
-        });
-        let stdoutBuffer = '';
-        let stderr = '';
-        let outputSize = 0;
-        let settled = false;
-        const fail = (error: Error) => {
-          if (settled) return;
-          settled = true;
-          child.kill();
-          reject(error);
-        };
-        const emitLine = (line: string) => {
-          if (!line.trim()) return true;
-          try {
-            onLine(line);
-            return true;
-          } catch (reason) {
-            fail(
-              reason instanceof Error
-                ? reason
-                : new Error(`invalid Work event: ${String(reason)}`),
-            );
-            return false;
-          }
-        };
-        child.stdout.on('data', (chunk) => {
-          const text = String(chunk);
-          outputSize += text.length;
-          if (outputSize > options.maxBuffer) {
-            fail(new Error('Work event stream exceeded maxBuffer'));
-            return;
-          }
-          stdoutBuffer += text;
-          const lines = stdoutBuffer.split(/\r?\n/u);
-          stdoutBuffer = lines.pop() ?? '';
-          for (const line of lines) {
-            if (!emitLine(line)) return;
-          }
-        });
-        child.stderr.on('data', (chunk) => {
-          stderr += String(chunk);
-        });
-        child.once('error', fail);
-        child.once('close', (code, signal) => {
-          if (settled) return;
-          if (!emitLine(stdoutBuffer)) return;
-          if (code !== 0) {
-            fail(
-              new Error(
-                stderr.trim() ||
-                  `Work event stream exited ${code ?? signal ?? 'unknown'}`,
-              ),
-            );
-            return;
-          }
-          settled = true;
-          resolve();
-        });
+      return streamCliEvents({
+        file,
+        args: cli.args(values),
+        env: tuiAgentSessionEnvironment(options.env),
+        maxBuffer: options.maxBuffer,
+        onLine,
+        label: 'Work',
       });
     },
   });
@@ -2190,7 +2128,6 @@ function ProductHost({
       />
     );
   }
-
   const resultCount =
     control.mode === 'commands'
       ? quickCommands.length
